Sprint 3

+++ dashboard/student.py
@@ -1,10 +1,12 @@
 import json
 from bson.json_util import dumps
 from flask import Blueprint, request
-from .db import insert_student, get_student, getAll_student, delete_student, update_student
+
+from .db import *
+
 from bson.objectid import ObjectId
 from flask_restful import Api, Resource
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import *
 
 bp = Blueprint('students',__name__)
 api = Api(bp)
@@ -17,35 +19,42 @@ class Students(Resource):
     
     @jwt_required()
     def post(self):
-        req = request.form
-        data = {
-            'nama' : req['nama'],
-            'jenis_kelamin' : req['jenis_kelamin'], 
-            'nisn' : req['nisn'], 
-            'nik' : req['nik'], 
-            'no_kk' : req['no_kk'], 
-            'tingkat_kelas' : req['tingkat_kelas'], 
-            'id_rombongan_belajar' : int(req['id_rombongan_belajar']), 
-            'tanggal_masuk' :req['tanggal_masuk'],
-            'tanggal_lulus' :req['tanggal_lulus'],
-            'nomor_induk' : req['nomor_induk'],
-            'status' : req['status'],
-            'tinggi_badan' : int(req['tinggi_badan']),
-            'berat_badan' : int(req['berat_badan']),
-            'lingkar_kepala' : int(req['lingkar_kepala']),
-            'alergi' : req['alergi'],
-            'nama_ayah' : req['nama_ayah'],
-            'nama_ibu' : req['nama_ibu'],
-            'pekerjaan_ayah' : req['pekerjaan_ayah'],
-            'pekerjaan_ibu' : req['pekerjaan_ibu']
-        }
-        insert_student(data)
-        return {'Success': True}
+        email = get_jwt_identity()
+        userDetail = get_user({"email":email})
+        if userDetail['is_admin']:
+            req = request.form
+            data = {
+                'nama' : req['nama'],
+                'jenis_kelamin' : req['jenis_kelamin'], 
+                'nisn' : req['nisn'], 
+                'nik' : req['nik'], 
+                'no_kk' : req['no_kk'], 
+                'tingkat_kelas' : req['tingkat_kelas'], 
+                'tahun_ajaran' : req['tahun_ajaran'],
+                'tanggal_masuk' :req['tanggal_masuk'],
+                'tanggal_lulus' :req['tanggal_lulus'],
+                'nomor_induk' : req['nomor_induk'],
+                'status' : req['status'],
+                'tinggi_badan' : int(req['tinggi_badan']),
+                'berat_badan' : int(req['berat_badan']),
+                'lingkar_kepala' : int(req['lingkar_kepala']),
+                'alergi' : req['alergi'],
+                'nama_ayah' : req['nama_ayah'],
+                'nama_ibu' : req['nama_ibu'],
+                'pekerjaan_ayah' : req['pekerjaan_ayah'],
+                'pekerjaan_ibu' : req['pekerjaan_ibu'],
+                'no_telp_ayah' : req['no_telp_ayah'],
+                'no_telp_ibu' : req['no_telp_ibu']
+            }
+            insert_student(data)
+            return {'Success': True}
+        else:
+            return {"Success" : False, "msg" : "Only admin can perform this action"}
 
 api.add_resource(Students, '/API/students')
 
 class Student(Resource):
-    @jwt_required()
+    # @jwt_required()
     def get(self, student_id):
         id = student_id
         ObjInstance = ObjectId(id)
@@ -55,45 +64,56 @@ class Student(Resource):
     
     @jwt_required()
     def put(self, student_id):
-        ObjInstance = ObjectId(student_id)
-        filter = {'_id':ObjInstance}
-        if (get_student(filter) is None):
-            return {"message":"ID not valid"}
-        else:
-            req = request.form
-            name = req.get("nama")
-            newVal = {
-                "$set":{
-                    'nama' : req['nama'],
-                    'jenis_kelamin' : req['jenis_kelamin'], 
-                    'nisn' : req['nisn'], 
-                    'nik' : req['nik'], 
-                    'no_kk' : req['no_kk'], 
-                    'tingkat_kelas' : req['tingkat_kelas'], 
-                    'id_rombongan_belajar' : int(req['id_rombongan_belajar']), 
-                    'tanggal_masuk' :req['tanggal_masuk'],
-                    'tanggal_lulus' :req['tanggal_lulus'],
-                    'nomor_induk' : req['nomor_induk'],
-                    'status' : req['status'],
-                    'tinggi_badan' : int(req['tinggi_badan']),
-                    'berat_badan' : int(req['berat_badan']),
-                    'lingkar_kepala' : int(req['lingkar_kepala']),
-                    'alergi' : req['alergi'],
-                    'nama_ayah' : req['nama_ayah'],
-                    'nama_ibu' : req['nama_ibu'],
-                    'pekerjaan_ayah' : req['pekerjaan_ayah'],
-                    'pekerjaan_ibu' : req['pekerjaan_ibu'],
+        email = get_jwt_identity()
+        userDetail = get_user({"email":email})
+        if userDetail['is_admin']:
+            ObjInstance = ObjectId(student_id)
+            filter = {'_id':ObjInstance}
+            if (get_student(filter) is None):
+                return {"message":"ID not valid"}
+            else:
+                req = request.form
+                newVal = {
+                    "$set":{
+                        'nama' : req['nama'],
+                        'jenis_kelamin' : req['jenis_kelamin'], 
+                        'nisn' : req['nisn'], 
+                        'nik' : req['nik'], 
+                        'no_kk' : req['no_kk'], 
+                        'tingkat_kelas' : req['tingkat_kelas'], 
+                        'tanggal_masuk' :req['tanggal_masuk'],
+                        'tanggal_lulus' :req['tanggal_lulus'],
+                        'nomor_induk' : req['nomor_induk'],
+                        'status' : req['status'],
+                        'tinggi_badan' : int(req['tinggi_badan']),
+                        'berat_badan' : int(req['berat_badan']),
+                        'lingkar_kepala' : int(req['lingkar_kepala']),
+                        'alergi' : req['alergi'],
+                        'nama_ayah' : req['nama_ayah'],
+                        'nama_ibu' : req['nama_ibu'],
+                        'pekerjaan_ayah' : req['pekerjaan_ayah'],
+                        'pekerjaan_ibu' : req['pekerjaan_ibu'],
+                        'no_telp_ayah' : req['no_telp_ayah'],
+                        'no_telp_ibu' : req['no_telp_ibu']
+                    }
                 }
-            }
 
-            update_student(filter, newVal)
-            return {"success":True}
+                update_student(filter, newVal)
+                return {"success":True}
+        else:
+            return {"Success" : False, "msg" : "Only admin can perform this action"}
+            
 
     @jwt_required()
     def delete(self, student_id):
-        ObjInstance = ObjectId(student_id)
-        filter = {'_id':ObjInstance}
-        data = delete_student(filter)
-        return {"success":True}
+        email = get_jwt_identity()
+        userDetail = get_user({"email":email})
+        if userDetail['is_admin']:
+            ObjInstance = ObjectId(student_id)
+            filter = {'_id':ObjInstance}
+            data = delete_student(filter)
+            return {"success":True}
+        else:
+            return {"Success" : False, "msg" : "Only admin can perform this action"}
 
 api.add_resource(Student, '/API/students/<student_id>')
